requirements.txt
appdirs
babel
kivy
kivymd
numpy
pandas
pycryptodomex
requests
urllib3